Refactor repository search methods to use consistent paging parameters.

+++ solo-leveling-player-model/src/main/java/com/sleepkqq/sololeveling/player/model/entity/user/LeaderboardUser.java
@@ -12,5 +12,5 @@ public class LeaderboardUser {
 
   private final LeaderboardUserView user;
   private final Number score;
-  private final Integer position;
+  private final Long position;
 }
